getenv_.c (G77_getenv_0): Avoid signed/unsigned warning.
	* libF77/getenv_.c (G77_getenv_0): Avoid signed/unsigned warning.
	* libF77/system_.c (G77_system_0): Likewise.
	* libI77/open.c (f_open): Likewise.
	* libI77/rdfmt.c (rd_Z): Likewise.
	* libI77/uio.c (do_us, do_ud): Likewise.

From-SVN: r54166

+++ libf2c/libF77/getenv_.c
@@ -24,7 +24,7 @@ G77_getenv_0 (char *fname, char *value, ftnlen flen, ftnlen vlen)
 
   if (flen <= 0)
     goto add_blanks;
-  for (i = 0; i < sizeof (buf); i++)
+  for (i = 0; i < (integer) sizeof (buf); i++)
     {
       if (i == flen || (buf[i] = fname[i]) == ' ')
 	{
